Nilanjana's patch + Avoid saving-restoring the register.
>>> Nilanjana's patch.
1. As we instrumneted the syscall instruction, and saving and restoring of the
   register is happening. After syscall hander is done executing, we want to
   restore all the registers except rax, which was restored before.
2. Avoided, libraries getting loaded multiple times because of multiple paths.
3. Enabled recursive guard.

>>> Avoiding saving and restoring the register.
We do not want to save and restore the register
before and after the instrumentation point.
Because we are instrumenting syscall instruction
and if there is no syscall handler then there is
no point of saving and restoring.

This should be done in syscall wrapper, which
will make sure to save the callie-save-registers
before calling syscall-handler, and if it want to
redirect the syscall to kernel itself then no
need to save any registers because its kernel's
job to make sure execution come back in userspace
with original values.